Adds League Spartan font & restructures CSS for better readability
