+++ src/css/App.css
@@ -1,4 +1,3 @@
 * {
-    margin: 0;
     box-sizing: border-box;
 }
