+++ src/css/index.css
@@ -1,0 +1,5 @@
+@import './fonts.css';
+
+* {
+    font-family: 'League Spartan';
+}
